Rewards split into active roles (#342)
* Rewards split into active roles

* Adjustments

* Reducing Garden size and tweaking regexp in error codes

* Another prettier round

* Test workaround and size reduction 24.03

* Reduced Garden.sol contract and added an event

* prettier

+++ contracts/strategies/Strategy.sol
@@ -901,7 +901,7 @@ contract Strategy is ReentrancyGuard, IStrategy, Initializable {
                 IRewardsDistributor(IBabController(controller).rewardsDistributor());
             // Only if the Mining program started on time for this strategy
             rewardsDistributor.substractProtocolPrincipal(capitalAllocated);
-            strategyRewards = rewardsDistributor.getStrategyRewards(address(this)); // Must be zero in case the mining program didnt started on time
+            strategyRewards = uint256(rewardsDistributor.getStrategyRewards(address(this))); // Must be zero in case the mining program didnt started on time
         }
     }
